Get rid of confusing _id and _count columns; use uuid as primary key throughout.

+++ app/src/main/java/org/projectbuendia/client/models/Location.java
@@ -28,10 +28,10 @@ public final class Location extends Base<String> {
     public final String uuid;
     public final String parentUuid;
     public final String name;
-    public final int patientCount;
+    public final long patientCount;
 
     /** Creates an instance of {@link Location}. */
-    public Location(String uuid, String parentUuid, String name, int patientCount) {
+    public Location(String uuid, String parentUuid, String name, long patientCount) {
         this.uuid = uuid;
         this.parentUuid = parentUuid;
         this.name = name;
